app/core/config.py
"""Environment-backed configuration for backend infrastructure."""

from __future__ import annotations

import os

try:
    from dotenv import load_dotenv
except ImportError:  # pragma: no cover - handled gracefully for pre-install environments
    load_dotenv = None

if load_dotenv is not None:
    load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_SERVICE_ROLE_KEY = os.getenv("SUPABASE_SERVICE_ROLE_KEY")
