requirements.txt
fastapi==0.111.0
lightgbm==4.6.0
numpy==2.0.1
openpyxl==3.1.5
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.5.1
supabase==2.15.0
treeinterpreter==0.2.3
uvicorn==0.30.1
xgboost==3.0.0
